+++ src/main/java/redis/clients/jedis/commands/ListPipelineCommands.java
@@ -74,7 +74,7 @@ public interface ListPipelineCommands {
 
   Response<KeyValue<String, List<String>>> lmpop(ListDirection direction, int count, String... keys);
 
-  Response<KeyValue<String, List<String>>> blmpop(long timeout, ListDirection direction, String... keys);
+  Response<KeyValue<String, List<String>>> blmpop(double timeout, ListDirection direction, String... keys);
 
-  Response<KeyValue<String, List<String>>> blmpop(long timeout, ListDirection direction, int count, String... keys);
+  Response<KeyValue<String, List<String>>> blmpop(double timeout, ListDirection direction, int count, String... keys);
 }
